refactor: init command
